Introduce ReactionsManager to centralize reactivity logic.
